refactor preview provider as region-specific
